JNG-4250 Rename packages
