--- judo-psm-archetype-generator/src/main/java/hu/blackbelt/judo/psm/archetype/generator/URLResolver.java
@@ -1,8 +1,0 @@
-package hu.blackbelt.judo.psm.archetype.generator;
-
-import java.io.IOException;
-import java.net.URL;
-
-public interface URLResolver {
-    URL getResource(String location) throws IOException;
-}
