+++ judo-psm-generator-engine/src/main/java/hu/blackbelt/judo/psm/generator/engine/GeneratedFile.java
@@ -1,0 +1,16 @@
+package hu.blackbelt.judo.psm.generator.engine;
+
+import lombok.Getter;
+import lombok.Setter;
+
+/**
+ * It represents a generated file. It is used to write to directory or create a ZipStream.
+ */
+@Getter
+@Setter
+public class GeneratedFile {
+
+	private String path;
+	private byte[] content;
+
+}
