+++ judo-psm-generator-engine/src/main/java/hu/blackbelt/judo/psm/generator/engine/URLResolver.java
@@ -1,0 +1,8 @@
+package hu.blackbelt.judo.psm.generator.engine;
+
+import java.io.IOException;
+import java.net.URL;
+
+public interface URLResolver {
+    URL getResource(String location) throws IOException;
+}
