+++ judo-psm-generator-maven-plugin/src/main/resources/META-INF/m2e/lifecycle-mapping-metadata.xml
